requirements.txt
playwright
requests
python-dotenv
pymongo
